@@ -1,15 +1,16 @@
 {
   "name": "frontend",
-  "proxy": "http://127.0.0.1:5000",
+  "proxy": "http://127.0.0.1:4000",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.21.0",
+    "bootstrap": "^5.1.3",
     "history": "^4.10.1",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-bootstrap": "^1.6.4",
+    "react-dom": "^17.0.2",
     "react-paypal-button-v2": "^2.6.2",
     "react-redux": "^7.2.2",
     "react-router-bootstrap": "^0.25.0",
