@@ -1,5 +1,5 @@
-import React from 'react';
-const rrd = require('react-router-dom');
+import React from "react";
+const rrd = require("react-router-dom");
 // Just render plain div with its children
-rrd.BrowserRouter = ({children}) => <div>{children}</div>
+rrd.BrowserRouter = ({ children }) => <div>{children}</div>;
 module.exports = rrd;
