@@ -3,9 +3,9 @@ import mongoose from "mongoose";
 // Async because it returns a promise
 const connectDB = async () => {
   try {
-      // This takes Mongo uri, then second argument with set of options
+    // This takes Mongo uri, then second argument with set of options
     const conn = await mongoose.connect(process.env.MONGO_URI, {
-        // with current version of mongoose we have to add few options, otherwise we'll be getting some warnings.
+      // with current version of mongoose we have to add few options, otherwise we'll be getting some warnings.
       useUnifiedTopology: true,
       useNewUrlParser: true,
       useCreateIndex: true,
@@ -13,7 +13,7 @@ const connectDB = async () => {
 
     console.log(`MongoDB connected: ${conn.connection.host}`.cyan.underline);
   } catch (error) {
-      // if failed to connect, return an error message and exit with failure
+    // if failed to connect, return an error message and exit with failure
     console.error(`Error: ${error.message}`.red.underline.bold);
     process.exit(1);
   }
