@@ -33,12 +33,12 @@ const protect = asyncHandler(async (req, res, next) => {
 });
 
 const admin = (req, res, next) => {
-  if(req.user && req.user.isAdmin) {
-    next()
+  if (req.user && req.user.isAdmin) {
+    next();
   } else {
-    res.status(401)
-    throw new Error('Not authorized as an admin')
+    res.status(401);
+    throw new Error("Not authorized as an admin");
   }
-}
+};
 
 export { protect, admin };
