@@ -1,8 +1,8 @@
-import path from 'path';
+import path from "path";
 import express from "express";
 import dotenv from "dotenv";
 import colors from "colors";
-import morgan from 'morgan';
+import morgan from "morgan";
 import { notFound, errorHandler } from "./middleware/errorMiddleware.js";
 import connectDB from "./config/db.js";
 
@@ -10,7 +10,7 @@ import connectDB from "./config/db.js";
 import productRoutes from "./routes/productRoutes.js";
 import userRoutes from "./routes/userRoutes.js";
 import orderRoutes from "./routes/orderRoutes.js";
-import uploadRoutes from './routes/uploadRoutes.js';
+import uploadRoutes from "./routes/uploadRoutes.js";
 
 dotenv.config();
 
@@ -18,8 +18,8 @@ connectDB();
 
 const app = express();
 
-if(process.env.NODE_ENV === 'development') {
-  app.use(morgan('dev'))
+if (process.env.NODE_ENV === "development") {
+  app.use(morgan("dev"));
 }
 
 app.use(express.json());
@@ -36,26 +36,28 @@ app.get("/api/config/paypal", (req, res) =>
 );
 
 // making uploads folder static so it can get loaded in the browser
-const __dirname = path.resolve()
-app.use('/uploads', express.static(path.join(__dirname, '/uploads')))
+const __dirname = path.resolve();
+app.use("/uploads", express.static(path.join(__dirname, "/uploads")));
 
-if(process.env.NODE_ENV === 'production') {
+if (process.env.NODE_ENV === "production") {
   // we want to set frontend/build folder as our static folder
   // so we can directly access it and load index.html
-  app.use(express.static(path.join(__dirname, '/frontend/build')))
+  app.use(express.static(path.join(__dirname, "/frontend/build")));
 
   // * - anything that isnt any of the above routes
-  app.get('*', (req, res) => res.sendFile(path.resolve(__dirname, 'frontend', 'build', 'index.html')))  
+  app.get("*", (req, res) =>
+    res.sendFile(path.resolve(__dirname, "frontend", "build", "index.html"))
+  );
 } else {
-  app.get('/', (req, res) => {
-    res.send("API RUNNING...")
-  })
+  app.get("/", (req, res) => {
+    res.send("API RUNNING...");
+  });
 }
 
 app.use(notFound);
 app.use(errorHandler);
 
-const PORT = process.env.PORT || 5000;
+const PORT = process.env.PORT || 4000;
 
 app.listen(
   PORT,
