@@ -1,10 +1,9 @@
-import jwt from 'jsonwebtoken';
+import jwt from "jsonwebtoken";
 
 const generateToken = (id) => {
-    return jwt.sign({ id }, process.env.JWT_SECRET, {
-        // expiresIn: '15min' //expires in 15 minutes 
-        expiresIn: '7d' 
-    })
-}
+  return jwt.sign({ id }, process.env.JWT_SECRET, {
+    expiresIn: "15min", //expires in 15 minutes
+  });
+};
 
 export default generateToken;
