@@ -4,8 +4,8 @@ import "../../setupTests";
 import Breadcrumb from "../../components/Breadcrumb.component";
 
 describe("Breadcrumb tests", () => {
-    it("Renders <Breadcrumb /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<Breadcrumb />)).toMatchSnapshot();
-    });
+  it("Renders <Breadcrumb /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<Breadcrumb />)).toMatchSnapshot();
+  });
 });
