@@ -4,8 +4,8 @@ import "../../setupTests";
 import CheckoutSteps from "../../components/CheckoutSteps.component";
 
 describe("CheckoutSteps tests", () => {
-    it("Renders <CheckoutSteps /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<CheckoutSteps />)).toMatchSnapshot();
-    });
+  it("Renders <CheckoutSteps /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<CheckoutSteps />)).toMatchSnapshot();
+  });
 });
