@@ -5,8 +5,8 @@ import Paginate from "../../components/Paginate.component";
 jest.mock("../../components/Paginate.component");
 
 describe("Paginate tests", () => {
-    it("Renders <Paginate /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<Paginate />)).toMatchSnapshot();
-    });
+  it("Renders <Paginate /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<Paginate />)).toMatchSnapshot();
+  });
 });
