@@ -5,8 +5,8 @@ import Product from "../../components/Product.component";
 jest.mock("../../components/Product.component");
 
 describe("Product tests", () => {
-    it("Renders <Product /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<Product />)).toMatchSnapshot();
-    });
+  it("Renders <Product /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<Product />)).toMatchSnapshot();
+  });
 });
