@@ -5,8 +5,8 @@ import Rating from "../../components/Rating.component";
 jest.mock("../../components/Rating.component");
 
 describe("Rating tests", () => {
-    it("Renders <Rating /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<Rating />)).toMatchSnapshot();
-    });
+  it("Renders <Rating /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<Rating />)).toMatchSnapshot();
+  });
 });
