@@ -5,8 +5,8 @@ import SearchBox from "../../components/SearchBox.component";
 jest.mock("../../components/SearchBox.component");
 
 describe("SearchBox tests", () => {
-    it("Renders <SearchBox /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<SearchBox />)).toMatchSnapshot();
-    });
+  it("Renders <SearchBox /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<SearchBox />)).toMatchSnapshot();
+  });
 });
