@@ -5,8 +5,8 @@ import Spinner from "../../components/Spinner.component";
 jest.mock("../../components/Spinner.component");
 
 describe("Spinner tests", () => {
-    it("Renders <Spinner /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<Spinner />)).toMatchSnapshot();
-    });
+  it("Renders <Spinner /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<Spinner />)).toMatchSnapshot();
+  });
 });
