@@ -1,11 +1,11 @@
 import React from "react";
 import { shallow } from "enzyme";
 import "../../setupTests";
-import AboutPage from "../../pages/AboutPage.component"
+import AboutPage from "../../pages/AboutPage.component";
 
 describe("<AboutPage />", () => {
-    it("Renders <AboutPage /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<AboutPage />)).toMatchSnapshot();
-    });
+  it("Renders <AboutPage /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<AboutPage />)).toMatchSnapshot();
+  });
 });
