@@ -1,14 +1,13 @@
 import React from "react";
 import { shallow } from "enzyme";
 import "../../setupTests";
-import CartPage from "../../pages/CartPage.component"
-import {useDispatch, useSelector} from '../../pages/CartPage.component';
+import CartPage from "../../pages/CartPage.component";
+import { useDispatch, useSelector } from "../../pages/CartPage.component";
 jest.mock("../../pages/CartPage.component");
 
 describe("<CartPage />", () => {
-    it("Renders <CartPage /> component correctly", () => {
-      // shallow creates instance of a component
-      expect(shallow(<CartPage />)).toMatchSnapshot();
-    });
+  it("Renders <CartPage /> component correctly", () => {
+    // shallow creates instance of a component
+    expect(shallow(<CartPage />)).toMatchSnapshot();
+  });
 });
-
