@@ -1,14 +1,7 @@
 import React from "react";
 import { Route } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import {
-  Navbar,
-  Nav,
-  NavDropdown,
-  Form,
-  FormControl,
-  Button,
-} from "react-bootstrap";
+import { Nav, Navbar, NavDropdown } from "react-bootstrap";
 import { LinkContainer } from "react-router-bootstrap";
 import Logo from "../assets/images/logo_top.svg";
 import SearchBox from "./SearchBox.component";
@@ -45,21 +38,6 @@ const Header = () => {
         </LinkContainer>
         <Navbar.Toggle aria-controls="responsive-navbar-nav" />
         <Navbar.Collapse id="responsive-navbar-nav">
-          <Nav className="mr-auto">
-            <LinkContainer to="/">
-              <Nav.Link className="nav-link">Home</Nav.Link>
-            </LinkContainer>
-            <LinkContainer to="/shop">
-              <Nav.Link className="nav-link">Shop</Nav.Link>
-            </LinkContainer>
-            <LinkContainer to="/contact">
-              <Nav.Link className="nav-link">Contact</Nav.Link>
-            </LinkContainer>
-            <LinkContainer to="/about">
-              <Nav.Link className="nav-link">About</Nav.Link>
-            </LinkContainer>
-          </Nav>
-
           <Nav className="ml-auto">
             <Route render={({ history }) => <SearchBox history={history} />} />
             <LinkContainer to="/cart">
