@@ -1,27 +1,27 @@
-import React, { useEffect } from 'react'
-import { LinkContainer } from 'react-router-bootstrap'
-import { Table, Button } from 'react-bootstrap'
-import { useDispatch, useSelector } from 'react-redux'
-import Message from '../components/Message.component'
-import Spinner from '../components/Spinner.component'
-import { listOrders } from '../actions/orderActions'
+import React, { useEffect } from "react";
+import { LinkContainer } from "react-router-bootstrap";
+import { Table, Button } from "react-bootstrap";
+import { useDispatch, useSelector } from "react-redux";
+import Message from "../components/Message.component";
+import Spinner from "../components/Spinner.component";
+import { listOrders } from "../actions/orderActions";
 
 const OrderListPage = ({ history }) => {
-  const dispatch = useDispatch()
+  const dispatch = useDispatch();
 
-  const orderList = useSelector((state) => state.orderList)
-  const { loading, error, orders } = orderList
+  const orderList = useSelector((state) => state.orderList);
+  const { loading, error, orders } = orderList;
 
-  const userLogin = useSelector((state) => state.userLogin)
-  const { userInfo } = userLogin
+  const userLogin = useSelector((state) => state.userLogin);
+  const { userInfo } = userLogin;
 
   useEffect(() => {
     if (userInfo && userInfo.isAdmin) {
-      dispatch(listOrders())
+      dispatch(listOrders());
     } else {
-      history.push('/login')
+      history.push("/login");
     }
-  }, [dispatch, history, userInfo])
+  }, [dispatch, history, userInfo]);
 
   return (
     <>
@@ -29,9 +29,9 @@ const OrderListPage = ({ history }) => {
       {loading ? (
         <Spinner />
       ) : error ? (
-        <Message variant='danger'>{error}</Message>
+        <Message variant="danger">{error}</Message>
       ) : (
-        <Table striped bordered hover responsive className='table-sm'>
+        <Table striped bordered hover responsive className="table-sm">
           <thead>
             <tr>
               <th>ID</th>
@@ -54,22 +54,21 @@ const OrderListPage = ({ history }) => {
                   {order.isPaid ? (
                     order.paidAt.substring(0, 10)
                   ) : (
-                    <i className='fas fa-times' style={{ color: 'red' }}></i>
+                    <i className="fas fa-times" style={{ color: "red" }}></i>
                   )}
                 </td>
                 <td>
                   {order.isDelivered ? (
                     order.deliveredAt.substring(0, 10)
                   ) : (
-                    <i className='fas fa-times' style={{ color: 'red' }}></i>
+                    <i className="fas fa-times" style={{ color: "red" }}></i>
                   )}
                 </td>
                 <td>
                   <LinkContainer to={`/order/${order._id}`}>
-                    <Button variant='light' className="btn-custom-blue py-3">
+                    <Button variant="light" className="btn-custom-blue py-3">
                       Details
                     </Button>
-                    
                   </LinkContainer>
                 </td>
               </tr>
@@ -78,7 +77,7 @@ const OrderListPage = ({ history }) => {
         </Table>
       )}
     </>
-  )
-}
+  );
+};
 
-export default OrderListPage
+export default OrderListPage;
