@@ -6,10 +6,12 @@ configure({ adapter: new Adapter() });
 /**
  * fix: `matchMedia` not present, legacy browsers require a polyfill
  */
-global.matchMedia = global.matchMedia || function() {
+global.matchMedia =
+  global.matchMedia ||
+  function () {
     return {
-        matches : false,
-        addListener : function() {},
-        removeListener: function() {}
-    }
-  }
+      matches: false,
+      addListener: function () {},
+      removeListener: function () {},
+    };
+  };
